fix(clippy): layer 2 gate lints (vox-db/populi/cli-review/+4)
deprecated codex_legacy test allow, unsafe test-mod allows (populi mesh_env/
domain_router), needless_option_as_deref, needless_borrow, map_or->is_some_and,
unwrap_or_else->unwrap_or, &PathBuf->&Path, redundant cast, Iterator::last->
next_back, sort_by->sort_by_key, useless vec!, module_inception allow,
empty-line-after-doc, assert_eq-bool, comparison-to-empty, unused glob imports.

+++ crates/vox-cli-review/src/coderabbit/config.rs
@@ -96,7 +96,7 @@ pub fn load_from_dir(path: &Path) -> CodeRabbitConfig {
         .or_else(|| resolved_tier.map(|t| t.min_delay_between_prs_secs()));
     let max_files = cr
         .max_files_per_pr
-        .or_else(|| resolved_tier.map(|t| t.recommended_max_files_per_pr() as u32));
+        .or_else(|| resolved_tier.map(|t| t.recommended_max_files_per_pr()));
 
     CodeRabbitConfig {
         tier: cr.tier,
